+++ src/components/plan/addOrEditPlan/AddOrEditPlan.tsx
@@ -1,16 +1,12 @@
 'use client';
 
 import React, { useRef } from 'react';
-import { useForm } from 'react-hook-form';
 
-import { yupResolver } from '@hookform/resolvers/yup';
 import { useQuery } from '@tanstack/react-query';
 
 import { getAllPinsByIdAndDates } from '@/api/pins';
 import Loading from '@/components/common/loading/Loading';
-import { addPlanSchema } from '@/schema/planSchema';
 import { useModifyPlanStoreActions, useModifyPlanStoreState } from '@/store/modifyPlanStore';
-import { addCommas } from '@/utils/numberFormat';
 
 import PostPlanForm from '../../common/form/PostPlanForm';
 import PlanTopBar from '../planTopBar/PlanTopBar';
@@ -41,22 +37,6 @@ const AddOrEditPlan = ({ plan }: Props) => {
     enabled: plan !== undefined,
   });
 
-  const resolver = yupResolver(addPlanSchema);
-
-  const {
-    register,
-    handleSubmit,
-    setValue,
-    formState: { errors, isValid },
-  } = useForm<PlanContentsInputType>({
-    resolver,
-    mode: 'onChange',
-    defaultValues: {
-      title: plan?.title,
-      totalCost: plan?.total_cost,
-    },
-  });
-
   const handleSaveOrModifyBtnClick = () => {
     if (modifyState === 'readOnly') {
       setModify();
@@ -64,16 +44,7 @@ const AddOrEditPlan = ({ plan }: Props) => {
     }
     if (formRef.current) {
       formRef.current.dispatchEvent(new Event('submit', { cancelable: true, bubbles: true }));
-
-      setReadOnly();
     }
-  };
-
-  const onChangeCost = (e: React.ChangeEvent<HTMLInputElement>) => {
-    let val = e.target.value.replace(/\D/g, '');
-    if (val.length > 8) val = val.substring(0, 8);
-
-    setValue('totalCost', addCommas(+val));
   };
 
   React.useEffect(() => {
@@ -95,11 +66,8 @@ const AddOrEditPlan = ({ plan }: Props) => {
         plan={plan}
         originPins={data}
         readonly={modifyState !== 'modify'}
+        setReadOnly={setReadOnly}
         formRef={formRef}
-        handleSubmit={handleSubmit}
-        onChangeCost={onChangeCost}
-        register={register}
-        errors={errors}
       />
       {isLoading && <Loading full={true} />}
     </>
